Update on "Add flag for dynamo+ddp optimizations"
Add a flag that can be used to turn dynamo+ddp optimizations on. This
will be used to compare how dynamo+ddp performs with and without the
additional graph break strategy for improving dynamo+ddp
compute/communication overlap.

Differential Revision: [D39976005](https://our.internmc.facebook.com/intern/diff/D39976005)

[ghstack-poisoned]
